[bugfix] fixed error message
fixed error message ""Stop" can only be called on an active agent that has been placed on a NavMesh."

Cause: SetMovementAllowed() was called through event invocation on enemy death. This is not allowed since SetMovementAllowed() can only be called on an enemy on the nav mesh.
Added error logs to catch this issue in the future.